Added AES-GCM with skrypt key derivation.
AES encryption replaced normal hex conversion.

The message is now shared alongside the IV or nonce.
The salt is used to derive the key from a password, the key has to remain same encrypting and decrypting so a fixed salt is used (it didn't make sense to me to share it in the steganographied message). For optimal security it has to be replaced but both parts should know it. Instructions for changing it and more disclaimers will be added in the readme file.